ACTIN-1811: split-o-rama: evidences for other attributes requiring evidence database
